+++ regex/regcomp.c
@@ -1,4 +1,4 @@
-/* $Id: regcomp.c,v 1.7 1999/11/12 20:57:49 sas Exp $ */
+/* $Id: regcomp.c,v 1.8 2000/02/02 19:00:53 fmk Exp $ */
 
 #include <sys/types.h>
 #include <stdio.h>
@@ -1056,7 +1056,7 @@ freeset(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register cset *top = &p->g->sets[p->g->ncsets];
 	register size_t css = (size_t)p->g->csetsize;
 
@@ -1082,7 +1082,7 @@ register struct parse *p;
 register cset *cs;
 {
 	register uch h = cs->hash;
-	register int i;
+	register size_t i;
 	register cset *top = &p->g->sets[p->g->ncsets];
 	register cset *cs2;
 	register size_t css = (size_t)p->g->csetsize;
@@ -1115,7 +1115,7 @@ firstch(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register size_t css = (size_t)p->g->csetsize;
 
 	for (i = 0; i < css; i++)
@@ -1134,7 +1134,7 @@ nch(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register size_t css = (size_t)p->g->csetsize;
 	register int n = 0;
